Extract session details builder into helper function
We want to reuse this process of building details from the post object in multiple places.
